# api/API.yaml
openapi: 3.0.3
info:
  title: Car Marketplace API
  version: 1.0.0
  description: |
    - **Market Price Analysis**: Calculate the average price of cars based on brand, model, year, and location.
    - **Filtered Search**: Allow users to search for cars based on specific criteria(make, model, year, price range, mileage, fuel type, location).
    - **Geographical Market Insights**: Analyze car availability and price variations across different regions.
    - **Seller Management & Profiling**: Distinguish between professional dealers and private sellers.
    - **Buyer–Seller Communication**: Enable real-time messaging between the buyer and the seller of a specific listing.
    - **Visitor & User Registration**: Track and register platform visitors, registered users can save favourite listings.
    - **Auction Module**: Allow sellers to list cars under an auction format.
    - **Listing details and comparison**: Provide detailed information on each listing and allow users to compare multiple listings side by side based on price, mileage, year and location.
...
paths:
...
  "/api/market/average-price":
    get:
      tags:
      - Market Price Analysis
      summary: Get average market price
      operationId: getAverageMarketPrice
      parameters:
      - "$ref": "#/components/parameters/BrandQueryRequired"
      - "$ref": "#/components/parameters/ModelQueryRequired"
      - "$ref": "#/components/parameters/YearFromQuery"
      - "$ref": "#/components/parameters/YearToQuery"
      responses:
        '200':
          description: Average market price.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/AveragePriceResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
...
  "/api/users/preview":
    post:
      tags:
      - Users
      summary: Get user previews
      operationId: getUsersPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UsersPreviewRequest"
      responses:
        '200':
          description: User previews.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/UsersPreviewResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
  "/api/sellers/{seller_id}":
    get:
      tags:
      - Sellers
      summary: Get seller profile
      operationId: getSellerProfile
      parameters:
      - "$ref": "#/components/parameters/SellerIdPath"
      responses:
        '200':
          description: Seller profile.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SellerProfileResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '404': *2
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
  "/api/sellers/preview":
    post:
      tags:
      - Sellers
      summary: Get seller previews
      operationId: getSellersPreview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SellersPreviewRequest"
      responses:
        '200':
          description: Seller previews.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SellersPreviewResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502': *1
  "/api/listings":
    get:
      tags:
      - Listings
      summary: Get listings
      operationId: getListings
      parameters:
      - "$ref": "#/components/parameters/PageQuery"
      - "$ref": "#/components/parameters/PageSizeCamelQuery"
      - "$ref": "#/components/parameters/IncludeSoldQuery"
      responses:
        '200':
          description: Listings.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SearchResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502': *1
    post:
      tags:
      - Listings
      summary: Create a listing
      operationId: createListing
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateListingRequest"
      responses:
        '201':
          description: Listing created.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ListingDetailsResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '401':
          "$ref": "#/components/responses/UnauthorizedText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '409': *3
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502': *1
...
  "/api/listings/compare":
    get:
      tags:
      - Listings
      summary: Compare listings
      operationId: compareListings
      parameters:
      - "$ref": "#/components/parameters/IdsQuery"
      responses:
        '200':
          description: Listing comparison.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CompareListingsResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
  "/api/listings/{id}":
    get:
      tags:
      - Listings
      summary: Get listing details
      operationId: getCarById
      parameters:
      - "$ref": "#/components/parameters/ListingIdPlainPath"
      responses:
        '200':
          description: Listing details.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ListingDetailsResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
    put:
      tags:
      - Listings
      summary: Update a listing
      operationId: updateListing
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/ListingIdPlainPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateListingRequest"
      responses:
        '200':
          description: Listing updated.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ListingDetailsResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '401':
          "$ref": "#/components/responses/UnauthorizedText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '409':
          "$ref": "#/components/responses/ConflictText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
    delete:
      tags:
      - Listings
      summary: Delete a listing
      operationId: deleteListing
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/ListingIdPlainPath"
      responses:
        '204':
          description: Listing deleted.
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '401':
          "$ref": "#/components/responses/UnauthorizedText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  parameters:
...
    ListingIdPlainPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    SellerIdPath:
      name: seller_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    BrandQueryRequired:
      name: brand
      in: query
      required: true
      schema:
        type: string
...
    IdsQuery:
      name: ids
      in: query
      required: false
      description: Comma-separated listing IDs. Invalid numbers are parsed as 0 by
        the gateway.
      schema:
        type: string
        example: 123
    IncludeSoldQuery:
      name: includeSold
      in: query
      required: false
      schema:
        type: boolean
...
    ModelQueryRequired:
      name: model
      in: query
      required: true
      schema:
        type: string
...
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 1
    PageSizeCamelQuery:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 20
...
    YearFromQuery:
      name: year_from
      in: query
      required: false
      schema:
        type: integer
        format: int32
    YearToQuery:
      name: year_to
      in: query
      required: false
      schema:
        type: integer
        format: int32
  responses:
    BadGatewayText:
      description: Upstream WebSocket connection failed.
      content:
        text/plain:
          schema:
            type: string
    BadRequestText:
      description: Bad request.
      content:
        text/plain:
          schema:
            type: string
    ConflictText:
      description: Conflict.
      content:
        text/plain:
          schema:
            type: string
...
    InternalServerErrorText:
      description: Internal server error.
      content:
        text/plain:
          schema:
            type: string
    MethodNotAllowedText:
      description: Method not allowed.
      content:
        text/plain:
          schema:
            type: string
            example: Method not allowed
    NotFoundText:
      description: Not found.
      content:
        text/plain:
          schema:
            type: string
            example: Not found
    UnauthorizedText:
      description: Unauthorized.
      content:
        text/plain:
          schema:
            type: string
            example: Unauthorized
  schemas:
...
    AveragePriceResponse:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        average_price:
          type: number
          format: double
        min_price:
          type: number
          format: double
        max_price:
          type: number
          format: double
        listing_count:
          type: integer
          format: int32
...
    CompareListingsResponse:
      type: object
      properties:
        listings:
          type: array
          items:
            "$ref": "#/components/schemas/ListingDetailsResponse"
...
    CreateListingRequest:
      type: object
      properties:
        vin:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
          format: int32
        price:
          type: number
          format: double
        mileage:
          type: integer
          format: int32
        city:
          type: string
        district:
          type: string
        state:
          type: string
        country:
          type: string
        fuel_type:
          type: string
        body_class:
          type: string
        drive_type:
          type: string
        transmission:
          type: string
        trim:
          type: string
        color:
          type: string
        dealer_id:
          type: integer
          format: int64
          readOnly: true
          description: Overwritten from the authenticated user by the gateway.
        is_new:
          type: boolean
        is_sold:
          type: boolean
...
    ListingDetailsResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        make:
          type: string
        model:
          type: string
        year:
          type: integer
          format: int32
        price:
          type: number
          format: double
        mileage:
          type: integer
          format: int32
        city:
          type: string
        district:
          type: string
        state:
          type: string
        country:
          type: string
        fuel_type:
          type: string
        trim:
          type: string
        transmission:
          type: string
        color:
          type: string
        seller_type:
          type: string
        description:
          type: string
        listed_at:
          type: string
        images:
          type: array
          items:
            type: string
        is_sold:
          type: boolean
    ListingSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        make:
          type: string
        model:
          type: string
        year:
          type: integer
          format: int32
        price:
          type: number
          format: float
        mileage:
          type: integer
          format: int32
        fuel_type:
          type: string
        body_class:
          type: string
        drive_type:
          type: string
        transmission:
          type: string
        is_new:
          type: boolean
        city:
          type: string
        district:
          type: string
        state:
          type: string
        country:
          type: string
        last_seen:
          type: string
        is_sold:
          type: boolean
        dealer_id:
          type: integer
          format: int32
...
    SearchResponse:
      type: object
      properties:
        total:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
        listings:
          type: array
          items:
            "$ref": "#/components/schemas/ListingSummary"
    SellerPreview:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    SellerProfileResponse:
      type: object
      properties:
        seller_id:
          type: integer
          format: int64
        name:
          type: string
        seller_type:
          type: string
        contact_info:
          type: string
        rating:
          type: number
          format: double
    SellersPreviewRequest:
      type: object
      properties:
        seller_ids:
          type: array
          items:
            type: integer
            format: int64
    SellersPreviewResponse:
      type: object
      properties:
        sellers:
          type: array
          items:
            "$ref": "#/components/schemas/SellerPreview"
...
    UpdateListingRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: Overwritten from the path by the gateway.
        vin:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
          format: int32
        price:
          type: number
          format: double
        mileage:
          type: integer
          format: int32
        city:
          type: string
        district:
          type: string
        state:
          type: string
        country:
          type: string
        fuel_type:
          type: string
        body_class:
          type: string
        drive_type:
          type: string
        transmission:
          type: string
        trim:
          type: string
        color:
          type: string
        dealer_id:
          type: integer
          format: int64
          readOnly: true
          description: Overwritten from the authenticated user by the gateway.
        is_new:
          type: boolean
    UserPreview:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    UsersPreviewRequest:
      type: object
      properties:
        user_ids:
          type: array
          items:
            type: integer
            format: int64
    UsersPreviewResponse:
      type: object
      properties:
        users:
          type: array
          items:
            "$ref": "#/components/schemas/UserPreview"